deployed_model_api/requirements.txt
Flask>=3.0.0
gunicorn>=21.2.0
joblib>=1.3.0
scikit-learn>=1.5.0
scikit-image>=0.22.0
Pillow>=10.0.0
numpy>=1.26.0
